requirements.txt
fastapi
uvicorn
motor
python-dotenv
passlib[bcrypt]